Parameterize link tests to steer AMQP I/O to another broker and path
Link tests have hard coded "amqp:guest:guest@127.0.0.1:5672" broker
and "q1" path. Changing these to point to another broker/path is a
problem.

This patch creates a TestTarget class that holds the broker
address and path. The default values are the same as before. TestTarget
can override those values with an environment variable in the form of
an Address URI.

  $ SET AMQPNETLITE_TESTTARGET=amqp://admin:password@example.com:9876/bigqueue
  $ mstest ...

Then each test file instantiates an instance of TestTarget and uses
that to locate the broker and what target to use for the test.

This patch has a few outstanding issues:

* I did not see an easy way to provide the override values on the mstest
command line. Good self tests are immune to environment variable settings.
The current solution is a compromise to allow expanded utility in a
simple form. Improvements are welcome.

* Hard coded localhost addresses, ports, and path values of "q1" still exist.
This patch does not try to solve them all.